Move Command.hpp
