+++ include/Utils/DataContainer.hpp
@@ -11,6 +11,7 @@
 #include "glm/vec2.hpp"
 #include "glm/vec3.hpp"
 #include "glm/vec4.hpp"
+#include "utils/Command.hpp"
 
 enum DataType {
     Float,
