--- src/Utils/Dictionary.cpp
@@ -1,5 +1,0 @@
-//
-// Created by Iannis on 6/29/2022.
-//
-
-#include "Utils/Dictionary.hpp"
